demo/requirements.txt
gradio_agentchatbot
git+https://github.com/huggingface/transformers.git#egg=transformers[agents]
langchain
langchain-community
sentence-transformers
faiss-cpu
serpapi
google-search-results